# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Friendtrip",
        "description": "Friendtrip api document (internal use only)",
        "contact": {
            "email": "chaniwat.meranote@gmail.com"
        },
        "version": "0.0.1"
    },
    "host": "friendtrip.dev",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/authentication": {
            "get": {
                "tags": [
                    "authentication"
                ],
                "summary": "Get authenticate user information",
                "description": "Get user information of given token",
                "operationId": "getAuthUserInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string",
                        "default": "Bearer "
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    }
                }
            },
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Authentication (Request token)",
                "description": "Request token for user auth needed APIs",
                "operationId": "authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User identity",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AuthenticationInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request token",
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    "401": {
                        "description": "Invalid credential",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AuthenticationInfo": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "",
                    "type": "string"
                },
                "password": {
                    "description": "",
                    "type": "string"
                },
                "get_info": {
                    "description": "",
                    "type": "boolean",
                    "default": false
                }
            },
            "type": "object"
        },
...
        "Token": {
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "",
                    "type": "string"
                },
                "user": {
                    "description": "",
                    "$ref": "#/definitions/User"
                }
            },
            "type": "object"
        },
...
        "User": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/UserBody"
                }
            ]
        },
        "UserBody": {
            "required": [
                "email",
                "first_name",
                "last_name",
                "display_name",
                "birthdate",
                "gender",
                "religion"
            ],
            "properties": {
                "email": {
                    "description": "",
                    "type": "string"
                },
                "first_name": {
                    "description": "",
                    "type": "string"
                },
                "last_name": {
                    "description": "",
                    "type": "string"
                },
                "display_name": {
                    "description": "",
                    "type": "string"
                },
                "birthdate": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "gender": {
                    "description": "",
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE"
                    ]
                },
                "religion": {
                    "description": "",
                    "type": "string"
                },
                "phone": {
                    "description": "",
                    "type": "string",
                    "maxLength": 10
                }
            },
            "type": "object"
        }
    },
...
}